requirements.txt
paramiko==3.5.0
python-dotenv==1.0.1
rocketchat_API==1.35.0
